# storage/api-docs/api-docs-v1.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Technique - HelloCSE",
        "description": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Test Technique v1"
        }
    ],
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Register a new administrator",
                "description": "Register a new administrator and return an access token and a refresh token.",
                "operationId": "d43ecdfd17aca209ecd9b0a8b370e2c0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful registration"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Log in as administrator",
                "description": "Log in an administrator and return an access token and a refresh token.",
                "operationId": "be2e78e5f93182d4403f839cc516d312",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            }
        },
        "/api/auth/refresh-token": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Refresh the access token",
                "description": "Refresh and return the access token.",
                "operationId": "58f1e2cba1b045b98578442e6260becf",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "refresh_token": []
                    }
                ]
            }
        },
        "/api/v1/profiles/{profile}/comments": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Store a new comment.",
                "description": "Store a new comment for the given profile.",
                "operationId": "0f949f23fb5fedde04114bf5020aeb91",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "The unique identifier of the profile",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCommentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful response: Comment created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/api/v1/profiles": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Get a list of active profiles.",
                "description": "Get a list of active profiles with pagination.",
                "operationId": "f301792bb5be5f7d940e04c16c077ab3",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of profiles per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 5
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Current page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of profiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProfileCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            },
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Store a new profile.",
                "description": "Store a new profile with the given data.",
                "operationId": "7cad69783ae4b4ad2f55eb818a6cae62",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profile created successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/api/v1/profiles/{profile}": {
            "put": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Update a profile.",
                "description": "Update the profile with the given data.",
                "operationId": "522545903b5825204314f1326a3a8cb1",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "The ID of the profile",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Delete a profile.",
                "description": "Delete the profile with the given ID.",
                "operationId": "7144b28409bed3f33d064b999499f213",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "The ID of the profile",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Profile deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "title": "Login Request",
                "description": "Request structure for administrator login",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "The email address of the administrator",
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "password": {
                        "description": "The password for the administrator",
                        "type": "string",
                        "format": "password",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "title": "Register Request",
                "description": "Request structure for administrator register",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "The email address of the administrator",
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "password": {
                        "description": "The password for the administrator",
                        "type": "string",
                        "format": "password",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "StoreCommentRequest": {
                "title": "Store Comment Request",
                "description": "Store Comment request body data",
                "required": [
                    "content"
                ],
                "properties": {
                    "content": {
                        "description": "The content of the comment",
                        "type": "string",
                        "example": "This is a comment"
                    }
                },
                "type": "object"
            },
            "StoreProfileRequest": {
                "title": "Store Profile Request",
                "description": "Request structure for storing a new profile.",
                "required": [
                    "last_name",
                    "first_name",
                    "image",
                    "status"
                ],
                "properties": {
                    "last_name": {
                        "description": "The last name of the profile",
                        "type": "string",
                        "example": "Doe"
                    },
                    "first_name": {
                        "description": "The first name of the profile",
                        "type": "string",
                        "example": "John"
                    },
                    "image": {
                        "description": "The image of the profile",
                        "type": "string",
                        "format": "base64"
                    },
                    "status": {
                        "description": "The status of the profile",
                        "type": "string",
                        "enum": [
                            "inactive",
                            "pending",
                            "active"
                        ]
                    }
                },
                "type": "object"
            },
            "UpdateProfileRequest": {
                "title": "Update Profile Request",
                "description": "Request structure for update a profile.",
                "properties": {
                    "last_name": {
                        "description": "The last name of the profile",
                        "type": "string",
                        "example": "Doe"
                    },
                    "first_name": {
                        "description": "The first name of the profile",
                        "type": "string",
                        "example": "John"
                    },
                    "image": {
                        "description": "The image of the profile",
                        "type": "string",
                        "format": "base64"
                    },
                    "status": {
                        "description": "The status of the profile",
                        "type": "string",
                        "enum": [
                            "inactive",
                            "pending",
                            "active"
                        ]
                    }
                },
                "type": "object"
            },
            "CommentCollection": {
                "title": "Comment Collection",
                "description": "Comment collection",
                "properties": {
                    "data": {
                        "description": "The list of comments",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CommentResource"
                        }
                    }
                },
                "type": "object"
            },
            "CommentResource": {
                "title": "Comment Resource",
                "description": "Comment resource",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the profile",
                        "type": "integer"
                    },
                    "profile_id": {
                        "description": "The unique identifier of the profile",
                        "type": "integer"
                    },
                    "administrator_id": {
                        "description": "The unique identifier of the administrator",
                        "type": "integer"
                    },
                    "content": {
                        "description": "The content of the comment",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "The date and time the comment was created",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ProfileCollection": {
                "title": "Profile Collection",
                "description": "Profile collection",
                "properties": {
                    "data": {
                        "description": "The list of profiles",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProfileResource"
                        }
                    }
                },
                "type": "object"
            },
            "ProfileResource": {
                "title": "Profile Resource",
                "description": "Profile resource",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the profile",
                        "type": "integer"
                    },
                    "administrator_id": {
                        "description": "The unique identifier of the administrator",
                        "type": "integer"
                    },
                    "last_name": {
                        "description": "The last name of the profile",
                        "type": "string"
                    },
                    "first_name": {
                        "description": "The first name of the profile",
                        "type": "string"
                    },
                    "image": {
                        "description": "The image of the profile",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "The date and time of the profile creation",
                        "type": "string",
                        "format": "date-time"
                    },
                    "comments": {
                        "description": "The comments of the profile",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CommentResource"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "access_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer {token})",
                "name": "Authorization",
                "in": "header"
            },
            "refresh_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer {token})",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Authentification",
            "description": "Group of endpoints for authentification."
        },
        {
            "name": "Comments",
            "description": "Group of endpoints for managing comments."
        },
        {
            "name": "Profiles",
            "description": "Group of endpoints for managing profiles."
        }
    ]
}